+++ common/src/main/java/ddraig/net/entropica/block/StationaryBrassTelescopeBlock.java
@@ -1,6 +1,7 @@
 package ddraig.net.entropica.block;
 
 import com.mojang.serialization.MapCodec;
+import ddraig.net.entropica.block.entity.StationaryBrassTelescopeBlockEntity;
 import ddraig.net.entropica.client.gui.SkyLookingGlassScreen;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -11,12 +12,16 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.EntityBlock;
 import net.minecraft.world.level.block.HorizontalDirectionalBlock;
+import net.minecraft.world.level.block.RenderShape;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.phys.BlockHitResult;
+import org.jetbrains.annotations.Nullable;
 
-public class StationaryBrassTelescopeBlock extends HorizontalDirectionalBlock {
+public class StationaryBrassTelescopeBlock extends HorizontalDirectionalBlock implements EntityBlock {
     public static final MapCodec<StationaryBrassTelescopeBlock> CODEC = simpleCodec(StationaryBrassTelescopeBlock::new);
 
     public StationaryBrassTelescopeBlock(Properties properties) {
@@ -40,10 +45,21 @@ public class StationaryBrassTelescopeBlock extends HorizontalDirectionalBlock {
     }
 
     @Override
+    public RenderShape getRenderShape(BlockState state) {
+        return RenderShape.MODEL;
+    }
+
+    @Nullable
+    @Override
+    public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
+        return new StationaryBrassTelescopeBlockEntity(pos, state);
+    }
+
+    @Override
     protected InteractionResult useWithoutItem(BlockState state, Level level, BlockPos pos, Player player, BlockHitResult hitResult) {
         level.playSound(player, pos, SoundEvents.SPYGLASS_USE, SoundSource.BLOCKS, 1.0f, 0.8f);
         if (level.isClientSide()) {
-            SkyLookingGlassScreen.open();
+            SkyLookingGlassScreen.openForTelescope(pos);
         }
         return InteractionResult.SUCCESS;
     }